feat(post): Finaliza funcionalidade de pinar/fixar post ao banner - Adiciona metodo e rota para update de um post existente - Adiciona regra para validação do usuário ao editar status pinned - Adiciona metodo para obter post pelo ID - Adiciona testes  para novos cenarios

+++ XUnitBlog.Data/Repositories/UserRepository.cs
@@ -22,7 +22,7 @@ internal class UserRepository(BlogContext blogContext) : IUserRepository
         return await blogContext.Users.FirstOrDefaultAsync(u => u.Email == email);
     }
 
-    public async Task<User?> GetUserById(int userId)
+    public async Task<User?> GetUserById(long userId)
     {
         return await blogContext.Users.AsNoTracking().FirstOrDefaultAsync(u => u.Id == userId);
     }
